feat: #1325 Split submit app form section into small components
