+++ packages/marketplace/src/components/pages/developer-submit-app/index.ts
@@ -1,0 +1,2 @@
+import DeveloperSubmitApp from './developer-submit-app'
+export default DeveloperSubmitApp
